Set ASDF_DATA_DIR to a tmpdir in BATS
I had some bad luck trying to use BATS_TEST_TMPDIR in setup(). On my
system, the variable was never set in tests.

So we are piggybacking on the BATS_RUN_TMPDIR to make our own tempdirs
in a directory that will be deleted by bats, and relying upon bats to
handle the cleanup.